+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 0.3.0
 -----
 
-* Removed deletion on map.
+* Removed `delete` on map entry lenses.
 * `in` does nothing if the path does not exist (unlike `update-in`).
+* Added `by-key`, `indexed` and `conditionally` lenses.
+* `view-single` throws an error if there are no foci.
 
 0.2.0
 -----